LayoutTests/fast/forms/datalist-nonoption-child-expected.txt

Test for child elements of a datalist element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS datalist.firstChild.nodeName is "#text"
PASS datalist.firstChild.nodeName is "B"
FAIL datalist.firstChild.nodeName should be #text. Was DIV.
PASS successfullyParsed is true

TEST COMPLETE
